requirements-dashboard.txt
# Dashboard API only — minimal dependencies
fastapi>=0.110.0
mangum>=0.17.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
